fix: removed duplicate userService. whoops.
